pyproject.toml
[tool.poetry]
name = "youtube_dash"
version = "0.1.0"
description = ""
authors = ["pedroolins"]

[tool.poetry.dependencies]
python = ">=3.9, <3.11"
pandas = "^1.4.1"
google-cloud = "^0.34.0"
google-auth = "^2.6.0"
google-cloud-bigquery = "^2.34.2"
google-api-python-client = "^2.39.0"
python-dotenv = "^0.19.2"
pandas-gbq = "^0.17.3"
Flask = "^2.0.3"
gunicorn = "^20.1.0"
pymsteams = "^0.2.1"
pytz = "^2021.3"
retrying = "^1.3.3"
plotly = "^5.6.0"
pytest = "^7.1.1"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
